# apps/backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RLApp Backend API
  version: 1.0.0
  description: |
    Contrato HTTP documentado a partir del código fuente actual de `apps/backend/src`.

    Este documento describe la implementación real expuesta hoy por el backend, no el contrato TO-BE descrito en `docs/project/07-interfaces-and-contracts`.
    Cuando el código y la documentación objetivo no coinciden, este OpenAPI prioriza el comportamiento implementado y marca la desviación en la descripción del endpoint o del schema.

    Limitaciones relevantes del estado actual:
    - La API no implementa versionado por URL.
    - Swagger/OpenAPI en runtime solo se publica en `Development`.
    - Hay campos `required` aceptados por la API que hoy no son usados por la lógica de aplicación.
    - El formato de errores no está unificado; conviven errores de validación ASP.NET, errores inline del controller y `ProblemDetails` del middleware.
servers:
  - url: http://localhost:5094
    description: Perfil HTTP local definido en `launchSettings.json`
  - url: https://localhost:7059
    description: Perfil HTTPS local definido en `launchSettings.json`
...
paths:
...
  /api/cashier/call-next:
    post:
      tags:
        - Cashier
      summary: Llamar el siguiente paciente en caja
      description: |
        Obtiene el siguiente paciente elegible de la cola y lo marca como llamado a caja.

        Hallazgos importantes:
        - Requiere política `CashierOperations`.
        - `cashierStationId` es obligatorio en el DTO, pero hoy no se usa en la lógica.
        - La respuesta real devuelve `patientId`, `calledAt` y `queuePosition`; no devuelve `turnId`, `turnNumber` ni `cashierStationId`.
      operationId: callNextAtCashier
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeaderRequired"
        - $ref: "#/components/parameters/IdempotencyKeyHeaderDeclaredRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallNextAtCashierRequest"
            examples:
              validCallNext:
                value:
                  queueId: Q-2026-03-19-MAIN
                  cashierStationId: CASH-01
      responses:
        "200":
          description: Paciente llamado a caja.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientCallResult"
              examples:
                success:
                  value:
                    patientId: PAT-0045
                    calledAt: "2026-03-19T14:15:00Z"
                    queuePosition: 1
        "400":
          $ref: "#/components/responses/BadRequestCommandOrValidation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/cashier/validate-payment:
    post:
      tags:
        - Cashier
      summary: Validar pago de un paciente
      description: |
        Marca el pago de un paciente como validado.

        Hallazgos importantes:
        - Requiere política `CashierOperations`.
        - `turnId` y `paymentReference` son obligatorios en el request DTO, pero no se usan en el command ni en el handler.
        - El flujo opera por `queueId` + `patientId`.
      operationId: validatePayment
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeaderOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidatePaymentRequest"
            examples:
              validPayment:
                value:
                  turnId: TURN-00045
                  queueId: Q-2026-03-19-MAIN
                  patientId: PAT-0045
                  paymentReference: PAY-20260319-0001
                  validatedAmount: 85.5
      responses:
        "200":
          description: Pago validado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
              examples:
                success:
                  value:
                    success: true
                    message: Payment validated successfully
                    correlationId: CORR-pay-001
                    executedAt: "2026-03-19T14:25:00Z"
        "400":
          description: Error de negocio o validación.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/InlineCommandError"
                  - $ref: "#/components/schemas/ValidationProblemDetails"
              examples:
                invalidAmount:
                  value:
                    error: Invalid payment amount
                    correlationId: CORR-pay-001
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/cashier/mark-payment-pending:
    post:
      tags:
        - Cashier
      summary: Marcar pago como pendiente
      description: |
        Marca el pago de un paciente como pendiente.

        Hallazgos importantes:
        - Requiere política `CashierOperations`.
        - `turnId`, `reason` y `attemptNumber` son obligatorios en el request DTO, pero hoy no se usan en el handler.
      operationId: markPaymentPending
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeaderOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkPaymentPendingRequest"
            examples:
              pendingPayment:
                value:
                  turnId: TURN-00045
                  queueId: Q-2026-03-19-MAIN
                  patientId: PAT-0045
                  reason: Datáfono temporalmente caído
                  attemptNumber: 1
      responses:
        "200":
          description: Pago marcado como pendiente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
              examples:
                success:
                  value:
                    success: true
                    message: Payment marked as pending for patient PAT-0045
                    correlationId: CORR-pending-001
                    executedAt: "2026-03-19T14:28:00Z"
        "400":
          $ref: "#/components/responses/BadRequestCommandOrValidation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/cashier/mark-absent:
    post:
      tags:
        - Cashier
      summary: Marcar ausencia en caja
      description: |
        Marca un paciente como ausente durante el flujo de caja.

        Hallazgos importantes:
        - Requiere política `CashierOperations`.
        - `turnId` y `reason` son obligatorios en el DTO, pero hoy no se usan.
        - El handler inyecta internamente `ROOM-CASHIER` como identificador de ubicación.
      operationId: markAbsentAtCashier
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeaderOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CashierMarkAbsentRequest"
            examples:
              absentAtCashier:
                value:
                  turnId: TURN-00045
                  queueId: Q-2026-03-19-MAIN
                  patientId: PAT-0045
                  reason: No respondió al llamado
      responses:
        "200":
          description: Ausencia registrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
              examples:
                success:
                  value:
                    success: true
                    message: Patient PAT-0045 marked as absent
                    correlationId: CORR-absent-cash-001
                    executedAt: "2026-03-19T14:30:00Z"
        "400":
          $ref: "#/components/responses/BadRequestCommandOrValidation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer emitido por `POST /api/staff/auth/login`.

        Roles implementados actualmente:
        - Receptionist
        - Cashier
        - Doctor
        - Supervisor
        - Support
  parameters:
    CorrelationIdHeaderRequired:
      name: X-Correlation-Id
      in: header
      required: true
      description: |
        Identificador de trazabilidad. El controller lo espera de forma obligatoria en este endpoint.
        No existe una política uniforme de validación; la obligatoriedad depende de cómo el action method fue declarado.
      schema:
        type: string
        example: CORR-9d8818f8
    CorrelationIdHeaderOptional:
      name: X-Correlation-Id
      in: header
      required: false
      description: |
        Identificador de trazabilidad opcional. Si no se envía, el controller genera uno.
      schema:
        type: string
        example: CORR-9d8818f8
    IdempotencyKeyHeaderDeclaredRequired:
      name: X-Idempotency-Key
      in: header
      required: true
      description: |
        Header declarado por el contrato mutante y recibido por el controller.
        Hallazgo crítico: actualmente no existe persistencia ni replay real de idempotencia; el valor se ignora en la lógica de aplicación.
      schema:
        type: string
        example: IDEMP-20260319-0001
...
  responses:
    BadRequestCommandOrValidation:
      description: Error de negocio inline o error de validación ASP.NET.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/InlineCommandError"
              - $ref: "#/components/schemas/ValidationProblemDetails"
          examples:
            commandFailure:
              value:
                error: Queue not found
                correlationId: CORR-err-001
            validationFailure:
              value:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.1
                title: One or more validation errors occurred.
                status: 400
                traceId: 00-a1b2c3d4e5f6-0000000000000000-00
                errors:
                  queueId:
                    - The QueueId field is required.
    Unauthorized:
      description: |
        Token ausente, inválido o expirado.
        La implementación actual usa el challenge por defecto de ASP.NET Core JWT y puede devolver una respuesta sin body estructurado.
    Forbidden:
      description: |
        El token es válido, pero el rol no satisface la policy del endpoint.
        La implementación actual puede responder sin body estructurado.
    InternalServerError:
      description: Error no controlado manejado por `GlobalExceptionMiddleware`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            unexpectedError:
              value:
                title: An unexpected error occurred.
                status: 500
                detail: The server could not process the request.
                instance: /api/cashier/validate-payment
                traceId: 00-a1b2c3d4e5f6-0000000000000000-00
                correlationId: CORR-500-001
  schemas:
...
    CallNextAtCashierRequest:
      type: object
      required:
        - queueId
        - cashierStationId
      properties:
        queueId:
          type: string
          example: Q-2026-03-19-MAIN
        cashierStationId:
          type: string
          description: Campo requerido en el DTO, pero hoy no se usa en el handler.
          example: CASH-01
    ValidatePaymentRequest:
      type: object
      required:
        - turnId
        - queueId
        - patientId
        - paymentReference
        - validatedAmount
      properties:
        turnId:
          type: string
          description: Campo requerido, pero actualmente ignorado por el backend.
          example: TURN-00045
        queueId:
          type: string
          example: Q-2026-03-19-MAIN
        patientId:
          type: string
          example: PAT-0045
        paymentReference:
          type: string
          description: Campo requerido, pero actualmente ignorado por el backend.
          example: PAY-20260319-0001
        validatedAmount:
          type: number
          format: double
          example: 85.5
    MarkPaymentPendingRequest:
      type: object
      required:
        - turnId
        - queueId
        - patientId
        - reason
        - attemptNumber
      properties:
        turnId:
          type: string
          description: Campo requerido, pero actualmente ignorado por el backend.
          example: TURN-00045
        queueId:
          type: string
          example: Q-2026-03-19-MAIN
        patientId:
          type: string
          example: PAT-0045
        reason:
          type: string
          description: Campo requerido, pero actualmente ignorado por el backend.
          example: Datáfono temporalmente caído
        attemptNumber:
          type: integer
          description: Campo requerido, pero actualmente ignorado por el backend.
          minimum: 1
          example: 1
    CashierMarkAbsentRequest:
      type: object
      required:
        - turnId
        - queueId
        - patientId
        - reason
      properties:
        turnId:
          type: string
          description: Campo requerido, pero actualmente ignorado por el backend.
          example: TURN-00045
        queueId:
          type: string
          example: Q-2026-03-19-MAIN
        patientId:
          type: string
          example: PAT-0045
        reason:
          type: string
          description: Campo requerido, pero actualmente ignorado por el backend.
          example: No respondió al llamado
...
    CommandResult:
      type: object
      required:
        - success
        - message
        - correlationId
        - executedAt
      properties:
        success:
          type: boolean
          example: true
...
    PatientCallResult:
      type: object
      required:
        - patientId
        - calledAt
        - queuePosition
      properties:
        patientId:
          type: string
          example: PAT-0045
        calledAt:
          type: string
          format: date-time
          example: "2026-03-19T14:15:00Z"
        queuePosition:
          type: integer
          example: 1
...
    InlineCommandError:
      type: object
      required:
        - error
        - correlationId
      properties:
        error:
          type: string
          description: |
            Mensaje devuelto directamente por el controller ante `CommandResult.Failure`.
            Puede ser un código canónico (`AUTH_INVALID_CREDENTIALS`) o un texto libre (`Queue not found`).
          example: Queue not found
        correlationId:
          type: string
          example: CORR-err-001
    ProblemDetails:
      type: object
      required:
        - title
        - status
        - detail
        - instance
        - traceId
      properties:
        title:
          type: string
          example: An unexpected error occurred.
        status:
          type: integer
          example: 500
        detail:
          type: string
          example: The server could not process the request.
        instance:
          type: string
          example: /api/cashier/validate-payment
        traceId:
          type: string
          example: 00-a1b2c3d4e5f6-0000000000000000-00
        correlationId:
          type: string
          nullable: true
          example: CORR-500-001
    ValidationProblemDetails:
      type: object
      required:
        - title
        - status
        - errors
      properties:
        type:
          type: string
          nullable: true
          example: https://tools.ietf.org/html/rfc9110#section-15.5.1
        title:
          type: string
          example: One or more validation errors occurred.
        status:
          type: integer
          example: 400
        traceId:
          type: string
          nullable: true
          example: 00-a1b2c3d4e5f6-0000000000000000-00
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            queueId:
              - The QueueId field is required.
...
security:
  - bearerAuth: []